requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-cors-headers==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djongo==1.2.35
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==6.1.0
pkg-resources==0.0.0
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.9.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
